refactor: change LiveAppDrawer folder
