+++ apps/ledger-live-desktop/src/renderer/components/LiveAppDrawer.tsx
